Move artifacts experience to separate page
- Moved masonry grid layout and lightbox functionality from home page to /artifacts page
- Updated home page to display only avatar, description, and link to artifacts
- Kept bottom logo animation on home page
- Added "View artifacts →" link to navigate to artifacts page
- Maintained all interactive features (keyboard navigation, touch/swipe, slide animations) on artifacts page
