+++ pages/index.js
@@ -1,48 +1,35 @@
-import { useEffect, useState, useCallback } from "react";
+import { useEffect, useState } from "react";
 import Image from "next/image";
-import Masonry from "react-masonry-css";
 import styles from "../styles.module.css";
-import artifactStyles from "../artifacts.module.css";
-import { artifacts } from "../data/artifacts";
 import { LinkPreview } from "../components/ui/link-preview";
 
 function HomePage() {
   const [isLoaded, setIsLoaded] = useState(false);
-  const [selectedArtifact, setSelectedArtifact] = useState(null);
-  const [nextArtifact, setNextArtifact] = useState(null);
-  const [isClosing, setIsClosing] = useState(false);
-  const [touchStart, setTouchStart] = useState(null);
-  const [touchEnd, setTouchEnd] = useState(null);
-  const [slideDirection, setSlideDirection] = useState(null); // 'left' or 'right'
-  const [isTransitioning, setIsTransitioning] = useState(false);
   const [logoProgress, setLogoProgress] = useState(0);
   const [showSocials, setShowSocials] = useState(false);
   const [hasShownSocials, setHasShownSocials] = useState(false);
   const [isDarkMode, setIsDarkMode] = useState(false);
 
-  // Minimum swipe distance (in px)
-  const minSwipeDistance = 50;
-
   useEffect(() => {
     // Ensure we're at the top of the page
     window.scrollTo(0, 0);
-    
+
     // Add loaded class to body to allow scrolling after animation
     document.body.classList.add('loaded');
-    
+
     // Trigger animation after component mounts
     setIsLoaded(true);
-    
+
     // Detect color scheme preference
     const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
     setIsDarkMode(mediaQuery.matches);
-    
+
     const handleColorSchemeChange = (e) => {
       setIsDarkMode(e.matches);
     };
-    
+
     mediaQuery.addEventListener('change', handleColorSchemeChange);
-    
+
     return () => {
       mediaQuery.removeEventListener('change', handleColorSchemeChange);
     };
@@ -54,13 +41,13 @@ function HomePage() {
       const scrollY = window.scrollY;
       const windowHeight = window.innerHeight;
       const documentHeight = document.documentElement.scrollHeight;
-      
+
       // Calculate distance from bottom
       const distanceFromBottom = documentHeight - (scrollY + windowHeight);
-      
+
       // Start animation when within 500px of bottom
       const triggerDistance = 500;
-      
+
       if (distanceFromBottom <= triggerDistance) {
         // Calculate progress (0 to 1)
         const progress = Math.min(1, 1 - (distanceFromBottom / triggerDistance));
@@ -73,121 +60,11 @@ function HomePage() {
     window.addEventListener('scroll', handleLogoScroll);
     // Call once on mount to set initial state
     handleLogoScroll();
-    
+
     return () => {
       window.removeEventListener('scroll', handleLogoScroll);
     };
   }, []);
-
-  const closeLightbox = () => {
-    setIsClosing(true);
-    setTimeout(() => {
-      setSelectedArtifact(null);
-      setIsClosing(false);
-    }, 300); // Match animation duration
-  };
-
-  const navigateToPrevious = useCallback(() => {
-    if (selectedArtifact === null || isTransitioning) return;
-    const currentIndex = artifacts.findIndex(a => a.id === selectedArtifact.id);
-    const prevIndex = currentIndex > 0 ? currentIndex - 1 : artifacts.length - 1;
-    const prevArtifact = artifacts[prevIndex];
-    
-    setIsTransitioning(true);
-    setSlideDirection('right'); // Current slides out right, new slides in from left
-    setNextArtifact(prevArtifact);
-    
-    setTimeout(() => {
-      setSelectedArtifact(prevArtifact);
-      setNextArtifact(null);
-      setIsTransitioning(false);
-      setSlideDirection(null);
-    }, 300);
-  }, [selectedArtifact, isTransitioning]);
-
-  const navigateToNext = useCallback(() => {
-    if (selectedArtifact === null || isTransitioning) return;
-    const currentIndex = artifacts.findIndex(a => a.id === selectedArtifact.id);
-    const nextIndex = currentIndex < artifacts.length - 1 ? currentIndex + 1 : 0;
-    const nextArtifactItem = artifacts[nextIndex];
-    
-    setIsTransitioning(true);
-    setSlideDirection('left'); // Current slides out left, new slides in from right
-    setNextArtifact(nextArtifactItem);
-    
-    setTimeout(() => {
-      setSelectedArtifact(nextArtifactItem);
-      setNextArtifact(null);
-      setIsTransitioning(false);
-      setSlideDirection(null);
-    }, 300);
-  }, [selectedArtifact, isTransitioning]);
-
-  useEffect(() => {
-    const handleKeyDown = (e) => {
-      if (selectedArtifact === null) return;
-
-      if (e.key === 'Escape') {
-        closeLightbox();
-      } else if (e.key === 'ArrowLeft') {
-        e.preventDefault();
-        navigateToPrevious();
-      } else if (e.key === 'ArrowRight') {
-        e.preventDefault();
-        navigateToNext();
-      }
-    };
-
-    // Prevent body scroll when lightbox is open
-    if (selectedArtifact !== null) {
-      document.body.style.overflow = 'hidden';
-      document.body.style.height = '100vh';
-      document.body.style.height = '100dvh';
-    } else {
-      document.body.style.overflow = '';
-      document.body.style.height = '';
-    }
-
-    window.addEventListener('keydown', handleKeyDown);
-    
-    return () => {
-      window.removeEventListener('keydown', handleKeyDown);
-      // Clean up body styles
-      document.body.style.overflow = '';
-      document.body.style.height = '';
-    };
-  }, [selectedArtifact, navigateToPrevious, navigateToNext]);
-
-  const handleArtifactClick = (artifact) => {
-    setIsClosing(false);
-    setSelectedArtifact(artifact);
-    setNextArtifact(null);
-    setSlideDirection(null);
-    setIsTransitioning(false);
-  };
-
-  const onTouchStart = (e) => {
-    setTouchEnd(null);
-    setTouchStart(e.targetTouches[0].clientX);
-  };
-
-  const onTouchMove = (e) => {
-    setTouchEnd(e.targetTouches[0].clientX);
-  };
-
-  const onTouchEnd = () => {
-    if (!touchStart || !touchEnd) return;
-    
-    const distance = touchStart - touchEnd;
-    const isLeftSwipe = distance > minSwipeDistance;
-    const isRightSwipe = distance < -minSwipeDistance;
-    
-    if (isLeftSwipe) {
-      navigateToNext();
-    } else if (isRightSwipe) {
-      navigateToPrevious();
-    }
-  };
 
   return (
     <div className={`${styles.container} ${isLoaded ? styles.loaded : ""}`}>
@@ -285,126 +162,11 @@ function HomePage() {
         </p>
       </div>
 
-      {/* Artifacts Section */}
-      <div className={artifactStyles.artifactsSection}>
-        {(() => {
-          const sections = [];
-          let currentGrid = [];
-          let animationIndex = 0;
-
-          const breakpointColumnsObj = {
-            default: 3,
-            1200: 3,
-            768: 2,
-            500: 1
-          };
-
-          artifacts.forEach((artifact, index) => {
-            if (artifact.featured) {
-              // Render accumulated grid section if it has items
-              if (currentGrid.length > 0) {
-                sections.push(
-                  <Masonry
-                    key={`grid-${index}`}
-                    breakpointCols={breakpointColumnsObj}
-                    className={artifactStyles.masonryGrid}
-                    columnClassName={artifactStyles.masonryGridColumn}
-                  >
-                    {currentGrid.map((item) => (
-                      <div 
-                        key={item.artifact.id} 
-                        className={artifactStyles.artifactItem}
-                        style={{ animationDelay: `${item.animationIndex * 0.05}s` }}
-                      >
-                        <div 
-                          className={artifactStyles.imageWrapper}
-                          onClick={() => handleArtifactClick(item.artifact)}
-                        >
-                          <img
-                            src={item.artifact.image}
-                            alt={item.artifact.caption}
-                            className={artifactStyles.artifactImage}
-                            loading="lazy"
-                          />
-                        </div>
-                        <p className={artifactStyles.caption}>
-                          {item.artifact.caption}
-                        </p>
-                      </div>
-                    ))}
-                  </Masonry>
-                );
-                currentGrid = [];
-              }
-
-              // Render featured artifact
-              sections.push(
-                <div 
-                  key={`featured-${artifact.id}`} 
-                  className={artifactStyles.featuredArtifact}
-                  style={{ animationDelay: `${animationIndex * 0.05}s` }}
-                >
-                  <div 
-                    className={artifactStyles.featuredImageWrapper}
-                    onClick={() => handleArtifactClick(artifact)}
-                  >
-                    <img
-                      src={artifact.image}
-                      alt={artifact.caption}
-                      className={artifactStyles.featuredImage}
-                      loading="lazy"
-                    />
-                  </div>
-                  <p className={artifactStyles.featuredCaption}>
-                    {artifact.caption}
-                  </p>
-                </div>
-              );
-              animationIndex++;
-            } else {
-              // Add to current grid section
-              currentGrid.push({ artifact, animationIndex });
-              animationIndex++;
-            }
-          });
-
-          // Render any remaining grid items
-          if (currentGrid.length > 0) {
-            sections.push(
-              <Masonry
-                key={`grid-final`}
-                breakpointCols={breakpointColumnsObj}
-                className={artifactStyles.masonryGrid}
-                columnClassName={artifactStyles.masonryGridColumn}
-              >
-                {currentGrid.map((item) => (
-                  <div 
-                    key={item.artifact.id} 
-                    className={artifactStyles.artifactItem}
-                    style={{ animationDelay: `${item.animationIndex * 0.05}s` }}
-                  >
-                    <div 
-                      className={artifactStyles.imageWrapper}
-                      onClick={() => handleArtifactClick(item.artifact)}
-                    >
-                      <img
-                        src={item.artifact.image}
-                        alt={item.artifact.caption}
-                        className={artifactStyles.artifactImage}
-                        loading="lazy"
-                      />
-                    </div>
-                    <p className={artifactStyles.caption}>
-                      {item.artifact.caption}
-                    </p>
-                  </div>
-                ))}
-              </Masonry>
-            );
-          }
-
-          return sections;
-        })()}
+      {/* Artifacts Link */}
+      <div className={styles.artifactsLinkSection}>
+        <a href="/artifacts" className={styles.artifactsLink}>
+          View artifacts →
+        </a>
       </div>
 
       {/* Bottom Logo */}
@@ -417,7 +179,7 @@ function HomePage() {
             className={styles.logo}
             style={{
               opacity: logoProgress,
-              filter: `blur(${10 - (logoProgress * 10)}px) ${isDarkMode 
+              filter: `blur(${10 - (logoProgress * 10)}px) ${isDarkMode
                 ? 'brightness(0) saturate(100%) invert(29%) sepia(0%) saturate(0%) hue-rotate(180deg) brightness(95%) contrast(92%)'
                 : 'brightness(0) saturate(100%) invert(60%) sepia(1%) saturate(0%) hue-rotate(0deg) brightness(105%) contrast(92%)'
               }`,
@@ -426,91 +188,6 @@ function HomePage() {
           />
         </div>
       </div>
-
-      {/* Lightbox */}
-      {selectedArtifact && (
-        <div 
-          className={`${artifactStyles.lightbox} ${isClosing ? artifactStyles.lightboxClosing : ''}`} 
-          onClick={closeLightbox}
-          onTouchStart={onTouchStart}
-          onTouchMove={onTouchMove}
-          onTouchEnd={onTouchEnd}
-        >
-          <div className={artifactStyles.lightboxBackdrop}></div>
-          <div className={artifactStyles.counter}>
-            {artifacts.findIndex(a => a.id === selectedArtifact.id) + 1}/{artifacts.length}
-          </div>
-          <button className={artifactStyles.closeButton} onClick={closeLightbox} aria-label="Close">
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-              <line x1="18" y1="6" x2="6" y2="18"></line>
-              <line x1="6" y1="6" x2="18" y2="18"></line>
-            </svg>
-          </button>
-          <button className={artifactStyles.navButton} style={{ left: '20px' }} onClick={(e) => { e.stopPropagation(); navigateToPrevious(); }} aria-label="Previous">
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-              <polyline points="15 18 9 12 15 6"></polyline>
-            </svg>
-          </button>
-          <button className={artifactStyles.navButton} style={{ right: '20px' }} onClick={(e) => { e.stopPropagation(); navigateToNext(); }} aria-label="Next">
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-              <polyline points="9 18 15 12 9 6"></polyline>
-            </svg>
-          </button>
-          <div className={`${artifactStyles.lightboxContent} ${isClosing ? artifactStyles.lightboxContentClosing : ''}`} onClick={(e) => e.stopPropagation()}>
-            <div className={artifactStyles.lightboxImageWrapper} onClick={closeLightbox}>
-              {/* Current image - slides out */}
-              {selectedArtifact && (
-                <div 
-                  className={`${artifactStyles.lightboxImageContainer} ${
-                    isTransitioning && slideDirection === 'left'
-                      ? artifactStyles.slideOutLeft 
-                      : isTransitioning && slideDirection === 'right'
-                      ? artifactStyles.slideOutRight 
-                      : ''
-                  }`}
-                >
-                  <Image
-                    key={selectedArtifact.id}
-                    src={selectedArtifact.image}
-                    alt={selectedArtifact.caption}
-                    layout="fill"
-                    objectFit="contain"
-                    className={artifactStyles.lightboxImage}
-                    quality={90}
-                    priority
-                  />
-                </div>
-              )}
-              {/* Next image - slides in */}
-              {nextArtifact && (
-                <div 
-                  className={`${artifactStyles.lightboxImageContainer} ${
-                    slideDirection === 'left'
-                      ? artifactStyles.slideInRight
-                      : slideDirection === 'right'
-                      ? artifactStyles.slideInLeft
-                      : ''
-                  }`}
-                >
-                  <Image
-                    key={nextArtifact.id}
-                    src={nextArtifact.image}
-                    alt={nextArtifact.caption}
-                    layout="fill"
-                    objectFit="contain"
-                    className={artifactStyles.lightboxImage}
-                    quality={90}
-                    priority
-                  />
-                </div>
-              )}
-            </div>
-            <p className={artifactStyles.lightboxCaption}>
-              {(nextArtifact || selectedArtifact)?.caption}
-            </p>
-          </div>
-        </div>
-      )}
     </div>
   );
 }
